# backend/src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ReviewResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comment": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"propertyId",
					"rating",
					"comment",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateReviewRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comment": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"propertyId",
					"rating"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReviewRequest": {
				"properties": {
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comment": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReviewStatsResponse": {
				"properties": {
					"propertyId": {
						"type": "number",
						"format": "double"
					},
					"averageRating": {
						"type": "number",
						"format": "double"
					},
					"totalReviews": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"propertyId",
					"averageRating",
					"totalReviews"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaymentResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"bookingId": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"paymentMethod": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"paidAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"bookingId",
					"amount",
					"paymentMethod",
					"status",
					"paidAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePaymentRequest": {
				"properties": {
					"bookingId": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"paymentMethod": {
						"type": "string"
					}
				},
				"required": [
					"bookingId",
					"amount",
					"paymentMethod"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePaymentRequest": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"paid",
							"failed"
						]
					},
					"paidAt": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ActivityLogResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"userId": {
						"type": "number",
						"format": "double"
					},
					"action": {
						"type": "string"
					},
					"entityType": {
						"type": "string"
					},
					"entityId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"action",
					"entityType",
					"entityId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateActivityLogRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double"
					},
					"action": {
						"type": "string"
					},
					"entityType": {
						"type": "string"
					},
					"entityId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"userId",
					"action",
					"entityType",
					"entityId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ngekost-aja-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/reviews": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ReviewResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List reviews with filters",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "propertyId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minRating",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxRating",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Create a review",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReviewRequest"
							}
						}
					}
				}
			}
		},
		"/reviews/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Get review by ID",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Update review",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReviewRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Remove",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Delete review",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/reviews/property/{propertyId}/stats": {
			"get": {
				"operationId": "GetPropertyStats",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReviewStatsResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Get review statistics for a property",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "propertyId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/payments": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PaymentResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List payments with filters",
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "bookingId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"pending",
								"paid",
								"failed"
							]
						}
					},
					{
						"in": "query",
						"name": "paymentMethod",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"description": "Create a payment",
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentRequest"
							}
						}
					}
				}
			}
		},
		"/payments/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Get payment by ID",
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Update payment",
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePaymentRequest"
							}
						}
					}
				}
			}
		},
		"/payments/{id}/process": {
			"post": {
				"operationId": "Process",
				"responses": {
					"200": {
						"description": "Processed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Process payment (mark as paid)",
				"tags": [
					"Payment"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/activity-logs": {
			"get": {
				"operationId": "List",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ActivityLogResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "List activity logs with filters",
				"tags": [
					"ActivityLog"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "entityType",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "entityId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 50,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActivityLogResponse"
								}
							}
						}
					}
				},
				"description": "Create activity log",
				"tags": [
					"ActivityLog"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateActivityLogRequest"
							}
						}
					}
				}
			}
		},
		"/activity-logs/cleanup": {
			"delete": {
				"operationId": "Cleanup",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deletedCount": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deletedCount"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete old activity logs",
				"tags": [
					"ActivityLog"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "daysOld",
						"required": false,
						"schema": {
							"default": 90,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}